Introduce -a to analyze preexisting traces and -O to set trace file path
This commit allows to specify the path of the tracefile, which was already
possible by setting the GRAPROF_OUT env variable.
This commit also allows to not start tracing, which was dropped in
in the previous commit da2aa2a70. Instead of generating a trace a
pre-existing trace file will be used.

Signed-off-by: Stefan Beller <stefanbeller@googlemail.com>